Revert "move initializing the SDK to app main file."
